move elements, all the attributes and events to their own place
