+++ crates/elements/src/_docs/color_syntax.rs
@@ -1,0 +1,22 @@
+//! ### Color syntax
+//!
+//! The attributes that have colors as values can use the following syntax:
+//!
+//! #### Static colors
+//! - `rect`
+//! - `blue`
+//! - `green`
+//! - `yellow`
+//! - `black` (default for `color` attribute)
+//! - `gray`
+//! - `white` (default for `background` attribute)
+//! - `orange`
+//! - `transparent`
+//!
+//! #### rgb() / hsl()
+//!
+//! - With RGB: `rgb(150, 60, 20)`
+//! - With RGB and alpha: `rgb(150, 60, 20, 0.7)`
+//!     - Yyou can also use 0-255 for the alpha: `rgb(150, 60, 20, 70)`
+//! - With HSL: `hsl(28deg, 80%, 50%)`
+//! - With HSL and alpha: `hsl(28deg, 80%, 50%, 25%)`
