+++ crates/elements/src/_docs/elements/background.md
@@ -1,0 +1,16 @@
+### background
+
+Specify a color as the background of an element.
+
+You can learn about the syntax of this attribute [here](#color-syntax).
+
+### Example:
+
+```rust, no_run
+fn app(cx: Scope) -> Element {
+    render!(
+        rect {
+            background: "red"
+        }
+    )
+}
