+++ crates/elements/src/_docs/elements/decoration.md
@@ -1,0 +1,22 @@
+### decoration
+
+Specify the decoration in a text.
+
+Accpted values:
+
+- `underline`
+- `line-through`
+- `overline`
+
+Example:
+
+```rust, no_run
+fn app(cx: Scope) -> Element {
+    render!(
+        label {
+            decoration: "line-through",
+            "Hello, World!"
+        }
+    )
+}
+```
