+++ crates/elements/src/_docs/elements/decoration_style.md
@@ -1,0 +1,25 @@
+### decoration_style
+
+Specify the decoration's style in a text.
+
+Accpted values:
+
+- `solid` (default)
+- `double`
+- `dotted`
+- `dashed`
+- `wavy`
+
+Example:
+
+```rust, no_run
+fn app(cx: Scope) -> Element {
+    render!(
+        label {
+            decoration: "line-through",
+            decoration_style: "dotted",
+            "Hello, World!"
+        }
+    )
+}
+```
