+++ crates/elements/src/_docs/elements/direction.md
@@ -1,0 +1,27 @@
+### direction
+
+Control how the inner elements will be stacked, possible values are `vertical` (default) and `horizontal`.
+
+##### Usage
+
+```rust, no_run
+fn app(cx: Scope) -> Element {
+    render!(
+        rect {
+            width: "100%",
+            height: "100%",
+            direction: "vertical",
+            rect {
+                width: "100%",
+                height: "50%",
+                background: "red"
+            },
+            rect {
+                width: "100%",
+                height: "50%",
+                background: "green"
+            }
+        }
+    )
+}
+```
