+++ crates/elements/src/_docs/elements/font_family.md
@@ -1,0 +1,18 @@
+### font_family
+
+With the `font_family` you can specify what font do you want to use for the inner text.
+
+Limitation: Only fonts installed in the system are supported for now.
+
+Example:
+
+```rust, no_run
+fn app(cx: Scope) -> Element {
+    render!(
+        label {
+            font_family: "Inter",
+            "Hello, World!"
+        }
+    )
+}
+```
