+++ crates/elements/src/_docs/elements/font_size.md
@@ -1,0 +1,16 @@
+### font_size
+
+You can specify the size of the text using `font_size`.
+
+Example:
+
+```rust, no_run
+fn app(cx: Scope) -> Element {
+    render!(
+        label {
+            font_size: "50",
+            "Hellooooo!"
+        }
+    )
+}
+```
