+++ crates/elements/src/_docs/elements/font_weight.md
@@ -1,0 +1,41 @@
+### font_weight
+
+You can choose a weight for a text using the `font_weight` attribute.
+
+Accepted values:
+
+- `invisible`
+- `thin`
+- `extra-light`
+- `light`
+- `normal` (default)
+- `medium`
+- `semi-bold`
+- `bold`
+- `extra-bold`
+- `black`
+- `extra-black`
+- `50`
+- `100`
+- `200`
+- `300`
+- `400`
+- `500`
+- `600`
+- `700`
+- `800`
+- `900`
+- `950`
+
+Example:
+
+```rust, no_run
+fn app(cx: Scope) -> Element {
+    render!(
+        label {
+            font_weight: "bold",
+            "Hello, World!"
+        }
+    )
+}
+```
