+++ crates/elements/src/_docs/elements/min_width_min_height.md
@@ -1,0 +1,21 @@
+### min_width & min_height
+
+`rect` supports specifying a minimum width and height, this can be useful if you use it alongside a percentage for the target size.
+
+See syntax for [`Size Units`](#size-units).
+
+##### Usage
+
+```rust, no_run
+fn app(cx: Scope) -> Element {
+    render!(
+        rect {
+            background: "red",
+            min_width: "100",
+            min_height: "100",
+            width: "50%",
+            height: "50%",
+        }
+    )
+}
+```
