+++ crates/elements/src/_docs/elements/rotate.md
@@ -1,0 +1,19 @@
+ ### rotate
+
+ The `rotate` attribute let's you rotate an element.
+
+ Example:
+
+ ```rust, no_run
+ fn app(cx: Scope) -> Element {
+     render!(
+         label {
+             rotate: "180deg",
+             "Hello, World!"
+         }
+     )
+ }
+ ```
+
+
+ Compatible elements: all except [`text`](/guides/elements.html#paragraph-and-text).
