+++ crates/elements/src/_docs/elements/width_height.md
@@ -1,0 +1,19 @@
+## width and height
+
+Specify the width and height for the given element.
+
+See syntax in [`Size Units`](crate::_docs::size_unit).
+
+### Example:
+
+```rust, no_run
+fn app(cx: Scope) -> Element {
+    render!(
+        rect {
+            background: "red",
+            width: "15",
+            height: "50",
+        }
+    )
+}
+```
