+++ crates/elements/src/lib.rs
@@ -1,2 +1,14 @@
-pub mod elements;
+//! ## Elements and events
+//!
+//! - [Elements](crate::elements#structs)
+//! - [Events](crate::elements#functions)
+//!
+
+#[cfg(doc)]
+pub mod _docs;
+
+mod definitions;
+pub mod elements {
+    pub use crate::definitions::*;
+}
 pub mod events;
