--- crates/freya/src/_docs/effects.rs
@@ -1,25 +1,0 @@
-//! # Effects
-//!
-//! Learn how the effects attributes work.
-//!
-//! - [`rotate`](#rotate)
-//!
-//! ### rotate
-//!
-//! The `rotate` attribute let's you rotate an element.
-//!
-//! Example:
-//!
-//! ```rust, no_run
-//! fn app(cx: Scope) -> Element {
-//!     render!(
-//!         label {
-//!             rotate: "180deg",
-//!             "Hello, World!"
-//!         }
-//!     )
-//! }
-//! ```
-//!
-//!
-//! Compatible elements: all except [`text`](/guides/elements.html#paragraph-and-text).
